+++ src/main/java/techcourse/myblog/utils/model/ModelUtil.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.model;
+package techcourse.myblog.utils.model;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
