+++ src/main/java/techcourse/myblog/utils/session/SessionContext.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.session;
+package techcourse.myblog.utils.session;
 
 public class SessionContext {
     public static final String USER = "user";
